Move `uncompress_file` from `airflow.utils` to Hive provider
It was only used in S3 To Hive module. One less thing in the `utils` directory.
